The subject form should allow editing for created documents.
This allows one form to do both jobs without duplicating significant
amount of render and component logic. On the constructor method, a new
property of `isUpdatingForm` is initiated depending on if data is passed
onto the component. Depending on this, new buttons are chosen to be
shown or not. Also, submitting the form causes different methods to
either create or update, again depending on the situation.