Significant reworking of the timing schema for part time teachers.
The timing object now contains only the days the teacher is scheduled to
come, whereas previously all the days were kept with empty values. This
change makes it easier to validate and reason about.

For implementation, the TeacherTiming component's onChange event returns
a value that is valid according to the schema. However, it doesn't do
anything to ensure pairs of timing have been entered. To do this, the
object is looped over and any value that equals ['', ''] causes the
property to be deleted from the object.

Side note: onChange() now fires after all pending state changes have
been completed.